requirements.txt
Django>=5.0,<5.3
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter==25.1
weasyprint==65.1
django-crispy-forms>=2.0,<3.0
crispy-bootstrap5>=0.7,<1.0
whitenoise>=6.6.0,<7.0
dj-database-url>=2.1.0,<3.0
psycopg2-binary>=2.9.9,<3.0
gunicorn>=21.2.0,<22.0
